MYRIAD-148 Addressing review comments
