+++ myriad-scheduler/src/main/java/com/ebay/myriad/configuration/MyriadConfiguration.java
@@ -18,7 +18,6 @@
  */
 package com.ebay.myriad.configuration;
 
-import com.ebay.myriad.configuration.ServiceConfiguration.OptionalSerializerStr;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import org.codehaus.jackson.map.annotate.JsonSerialize;
 import com.google.common.base.Optional;
@@ -27,6 +26,8 @@ import com.google.common.base.Strings;
 import org.hibernate.validator.constraints.NotEmpty;
 
 import java.util.Map;
+
+import com.ebay.myriad.configuration.OptionalSerializer.OptionalSerializerString;
 
 /**
  * Myriad Configuration commonly defined in the YML file
@@ -105,11 +106,11 @@ public class MyriadConfiguration {
   private String frameworkRole;
 
   @JsonProperty
-  @JsonSerialize(using = OptionalSerializerStr.class)
+  @JsonSerialize(using = OptionalSerializerString.class)
   private String frameworkUser;
 
   @JsonProperty
-  @JsonSerialize(using = OptionalSerializerStr.class)
+  @JsonSerialize(using = OptionalSerializerString.class)
   private String frameworkSuperUser;
 
   @JsonProperty
